GOVSI-696: Break `ConfigurationService` dependency on `NotificationType`
This is needed to move `ConfigurationService` into `shared/`
